Add code to allow deleting language that has components

+++ src/main/java/com/oultoncollege/ComponentCatalog/controllers/LanguageController.java
@@ -76,7 +76,10 @@ public class LanguageController {
     @PostMapping("/delete/{id}")
     public String deleteLanguageSubmit(@PathVariable("id") int id, Model model) {
         Language language = languageService.getLanguage(id);
-        languageService.deleteLanguage(language);
+
+        if (language != null)
+            languageService.deleteLanguage(language);
+
         return "redirect:/language/";
     }
 }
